add some unit tests and reorganise crawl module
