+++ src/crawl/mod.rs
@@ -1,5 +1,7 @@
 mod core_types;
 mod crawler;
+mod net;
 
 pub use core_types::*;
 pub use crawler::*;
+pub use net::*;
